move common methods out of dom files
